+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_da.js
@@ -1,5 +1,6 @@
 {
     load: 'Henter data ...',
+    reload: 'Henter data ...',
     update: 'Opdaterer data ...',
     submit: 'Indsendelse data ...',
     save: 'Opdaterer data ...',