+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_de.js
@@ -1,5 +1,6 @@
 {
     load: 'Daten werden geladen ...',
+    reload: 'Daten werden geladen ...',
     update: 'Aktualisieren von Daten ...',
     submit: 'Übermittlung von Daten ...',
     save: 'Aktualisieren von Daten ...',