+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_en.js
@@ -1,5 +1,6 @@
 {
     load: 'Loading data ...',
+    reload: 'Loading data ...',
     update: 'Updating data...',
     submit: 'Submitting data ...',
     save: 'Updating data ...',