+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_nb.js
@@ -1,5 +1,6 @@
 {
     load: 'Laster data ...',
+    reload: 'Laster data ...',
     update: 'Oppdatere data ...',
     submit: 'Sende data ...',
     save: 'Oppdatere data ...',