+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_nl.js
@@ -1,5 +1,6 @@
 {
     load: 'Laden gegevens ...',
+    reload: 'Laden gegevens ...',
     update: 'Updatingdata',
     submit: 'Versturen gegevens ...',
     save: 'Bewaren gegevens ...',