+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_pt.js
@@ -1,5 +1,6 @@
 {
     load: 'Carregamento de dados ...',
+    reload: 'Carregamento de dados ...',
     update: 'Atualizando dados ...',
     submit: 'Enviando dados ...',
     save: 'Atualizando dados ...',