+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_sk.js
@@ -1,5 +1,6 @@
 {
     load: 'Načítanie dát ...',
+    reload: 'Načítanie dát ...',
     update: 'Aktualizácia dát ...',
     submit: 'Odoslanie dát ...',
     save: 'Aktualizácia dát ...',