+++ src/gallery-itsamodelsyncpromise/lang/gallery-itsamodelsyncpromise_sv.js
@@ -1,5 +1,6 @@
 {
     load: 'Laddar ...',
+    reload: 'Laddar ...',
     update: 'Uppdatering uppgifter ...',
     submit: 'Skickar uppgifter ...',
     save: 'Uppdatering uppgifter ...',